requirements.txt
flask>=3.0.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
networkx>=3.2.1
matplotlib>=3.8.2
pyvis>=0.3.2
